+++ packages/common/CHANGELOG.md
@@ -7,6 +7,9 @@
 - ConcordiumHdWallet methods now take the identity provider index as arguments.
 - Bumped @concordium/rust-bindings to 0.4.0.
 
+### Fixed
+
+ - Added missing `accountAddress` field to `AccountInfo` types.
 
 ## 3.0.0 2022-8-26